Implement project opening via 'Ctrl + O' shortcut
